initial the project
initial the  project and change the structure
