+++ src/client/src/interfaces/cart.d.ts
@@ -1,0 +1,9 @@
+export interface ICartItem {
+	id: number;
+	category: string;
+	description: string;
+	image: string;
+	price: number;
+	title: string;
+	amount: number;
+}
