+++ src/client/src/pages/main.styles.tsx
